move genomics tests to subfolder
