+++ tests/genomics/test_antismash_loader.py
@@ -3,7 +3,7 @@ from nplinker.genomics import BGC
 from nplinker.genomics import BGCLoaderBase
 from nplinker.genomics.antismash import AntismashBGCLoader
 from nplinker.genomics.antismash import parse_bgc_genbank
-from . import DATA_DIR
+from .. import DATA_DIR
 
 
 class TestAntismashBGCLoader:
